// src/bindings/generated.ts
export type PuzzleDatabaseInfo = { 
/**
 * The title of the puzzle database (derived from filename)
 */
title: string; 
/**
 * Description of the puzzle database (currently not populated)
 * TODO: Consider adding a way to store and retrieve database descriptions
 */
description: string; 
/**
 * Number of puzzles in the database
 */
puzzleCount: number; 
/**
 * Size of the database file in bytes
 */
storageSize: bigint; 
/**
 * Full path to the database file
 */
path: string }